refactor: move ShareQrCode to its own package
